app/requirements.txt
fastapi
uvicorn
jinja2
playwright
python-multipart
python-dotenv
azure-ai-contentunderstanding
azure-identity
azure-core